переименование и разбиение на точку входа и резделения логики для тестов
